DDC-1925 - squashing ticket tests into a single file
