--- tests/Doctrine/Tests/Models/DDC1925/DDC1925User.php
@@ -1,56 +1,0 @@
-<?php
-
-namespace Doctrine\Tests\Models\DDC1925;
-
-/**
- * @Table
- * @Entity
- */
-class DDC1925User
-{
-    /**
-     * @var integer
-     *
-     * @Column(name="id", type="integer")
-     * @Id
-     * @GeneratedValue(strategy="AUTO")
-     */
-    private $id;
-
-    /**
-     * @var string
-     *
-     * @Column(name="title", type="string", length=255)
-     */
-    private $title;
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set title
-     *
-     * @param string $title
-     */
-    public function setTitle($title)
-    {
-        $this->title = $title;
-    }
-
-    /**
-     * Get title
-     *
-     * @return string
-     */
-    public function getTitle()
-    {
-        return $this->title;
-    }
-}
